Refactor: move classes to more specific package
