+++ examples/gae-autoinstrument-example/.gitignore
@@ -1,2 +1,4 @@
 # Maven Build target folder
 target/
+# IDE specific file
+nbactions.xml
