--- examples/gae-autoinstrument-example/nbactions.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<actions>
-    <action>
-       <actionName>CUSTOM-appengine:run</actionName>
-        <displayName>appengine:run</displayName>
-        <goals>
-            <goal>appengine:run</goal>
-        </goals>
-    </action>
-    <action>
-        <actionName>CUSTOM-appengine:deploy</actionName>
-        <displayName>appengine:deploy</displayName>
-        <goals>
-            <goal>appengine:deploy</goal>
-        </goals>
-    </action>
-</actions>
