+++ examples/gae-autoinstrument-example/src/main/java/com/google/cloud/opentelemetry/example/appengine/HelloAppEngine.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.google;
+package com.google.cloud.opentelemetry.example.appengine;
 
 import com.google.appengine.api.utils.SystemProperty;
