+++ examples/gae-autoinstrument-example/src/main/java/com/google/cloud/opentelemetry/example/appengine/logging/enhancers/LogEnhancer.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.google.logging.enhancers;
+package com.google.cloud.opentelemetry.example.appengine.logging.enhancers;
 
 import com.google.cloud.logging.LogEntry;
 import com.google.cloud.logging.LoggingEnhancer;
