+++ examples/gae-autoinstrument-example/src/main/resources/logback.xml
@@ -4,7 +4,7 @@
       <level>INFO</level>
     </filter>
     <!-- This enhancer correlates logs with traces -->
-    <loggingEventEnhancer>com.google.logging.enhancers.LogTraceCorrelationEnhancer</loggingEventEnhancer>
+    <loggingEventEnhancer>com.google.cloud.opentelemetry.example.appengine.logging.enhancers.LogTraceCorrelationEnhancer</loggingEventEnhancer>
     <log>application.log</log> <!-- Optional : default java.log -->
     <resourceType>gae_app</resourceType> <!-- Optional : default: auto-detected, fallback: global -->
     <flushLevel>INFO</flushLevel> <!-- Optional : default ERROR -->
