+++ examples/gae-autoinstrument-example/src/main/webapp/WEB-INF/logging.properties
@@ -16,7 +16,7 @@
 # The following creates the console handler
 handlers = java.util.logging.ConsoleHandler
 # explicitly setting a logging handler for HelloAppEngine class
-com.google.HelloAppEngine.handlers=com.google.cloud.logging.LoggingHandler
+com.google.cloud.opentelemetry.example.appengine.HelloAppEngine.handlers=com.google.cloud.logging.LoggingHandler
 # Set the default logging level for the root logger
 .level = INFO
 
@@ -34,7 +34,7 @@ com.google.cloud.logging.LoggingHandler.flushLevel=ERROR
 # default : java.log
 com.google.cloud.logging.LoggingHandler.log=application_log
 #optional enhancers (to add additional fields, labels)
-com.google.cloud.logging.LoggingHandler.enhancers=com.google.logging.enhancers.LogEnhancer
+com.google.cloud.logging.LoggingHandler.enhancers=com.google.cloud.opentelemetry.example.appengine.logging.enhancers.LogEnhancer
 
 
 # FinalizableReferenceQueue tries to spin up a thread and fails.  This
