+++ examples/gae-autoinstrument-example/src/main/webapp/index.jsp
@@ -1,6 +1,6 @@
 <!DOCTYPE html>
 <%@ page contentType="text/html;charset=UTF-8" language="java" %>
-<%@ page import="com.google.HelloAppEngine" %>
+<%@ page import="com.google.cloud.opentelemetry.example.appengine.HelloAppEngine" %>
 <html>
 <head>
   <link href='//fonts.googleapis.com/css?family=Marmelad' rel='stylesheet' type='text/css'>
