+++ examples/gae-autoinstrument-example/src/test/java/com/google/cloud/opentelemetry/example/appengine/HelloAppEngineTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.google;
+package com.google.cloud.opentelemetry.example.appengine;
 
 import static com.google.common.truth.Truth.assertThat;
 import static org.mockito.Mockito.when;
